Remove translation
